django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1